Imporved domains page code, moved to folder
