+++ ui/app/rif/pages/index.js
@@ -1,0 +1,7 @@
+import DomainsScreen from './domainsPage/domainsPage'
+import PaymentsScreen from './paymentsPage/paymentsPage'
+
+export {
+    DomainsScreen,
+    PaymentsScreen,
+}
